Update: Fix `/` commands and fix `/stop` command

+++ message/commands.js
@@ -8,6 +8,10 @@ function commands() {
         const file = 'message/messageLogs/' + contact.id.user;
         const now = new Date();
         const calendar = date.format(now, 'dddd, MMMM DD YYYY HH:mm:ss.SSS [UTC]ZZ');
+        if (message.body.slice(0, 7) === '/start ') {
+            const arg = message.body.slice(7);
+            client.sendMessage(message.from, 'Error command argument `' + arg + '`\nPlease use /start to start chat');
+        }
         if (message.body === '/start') {
             if (fs.existsSync(file)) {
                 client.sendMessage(message.from, 'You are already started a chat');
@@ -25,27 +29,42 @@ function commands() {
         let iMax = 9999;
         let fileOld;
         const folderStop = 'message/messageLogs/stoppedChats/' + contact.id.user + '/';
+        if (message.body.slice(0, 6) === '/stop ') {
+            const arg = message.body.slice(6);
+            client.sendMessage(message.from, 'Error command argument `' + arg + '`\nPlease use /stop to stop chat');
+        }
         if (message.body === '/stop') {
-            if (!fs.existsSync(folderStop)) {
-                fs.mkdirSync(folderStop);
-            }
-            while (i < iMax) {
-                fileOld = folderStop + contact.id.user + '.' + i;
-                if (!fs.existsSync(fileOld)) {
-                    break;
+            if (!fs.existsSync(file)){
+                client.sendMessage(message.from, 'You did\'nt start the chat\nPlease use /start to start a chat');
+            } else {
+                if (!fs.existsSync(folderStop)) {
+                    fs.mkdirSync(folderStop);
                 }
-                i++;
+                while (i < iMax) {
+                    fileOld = folderStop + contact.id.user + '.' + i;
+                    if (!fs.existsSync(fileOld)) {
+                        break;
+                    }
+                    i++;
+                }
+                const oldMessages = JSON.parse(fs.readFileSync(file));
+                oldMessages[0].stopped_at = calendar;
+                const json = JSON.stringify(oldMessages, null, 4);
+                fs.writeFileSync(fileOld, json);
+                fs.rmSync(file);
+                client.sendMessage(message.from, 'Chat was stopped');
+                client.sendMessage(message.from, 'To start chat again please use /start again');
             }
-            const oldMessages = JSON.parse(fs.readFileSync(file));
-            oldMessages[0].stopped_at = calendar;
-            const json = JSON.stringify(oldMessages, null, 4);
-            fs.writeFileSync(fileOld, json);
-            fs.rmSync(file);
-            client.sendMessage(message.from, 'Chat was stopped');
-            client.sendMessage(message.from, 'To start chat again please use /start again');
         }
 
         let listHistory = [];
+        if (message.body.slice(0, 9) === '/history ') {
+            if (message.body !== '/history show' &&
+                message.body.slice(0, 14) !== '/history show ') {
+                const arg = message.body.slice(9);
+                client.sendMessage(message.from, 'Error command argument `' + arg + '`\nPlease use /history to show stopped chat history');
+            }
+        }
         if (message.body === '/history') {
             if (!fs.existsSync(folderStop)) {
                 client.sendMessage(message.from, 'You don\'t have any chat history!');
@@ -85,6 +104,7 @@ function commands() {
                     client.sendMessage(message.from, 'Use: /history\nTo see list of your chat history');
                 } else {
                     const parseHistory = JSON.parse(fs.readFileSync(fileHistory)).slice(2);
+                    const parseHistoryDate = JSON.parse(fs.readFileSync(fileHistory)).slice(0, 1);
                     let a = 0;
                     const chats = [];
                     while (a < parseHistory.length) {
@@ -99,7 +119,7 @@ function commands() {
                         chats.push(author + ': ' + msg);
                         a++;
                     }
-                    client.sendMessage(message.from, chats.join('\n\n====================\n\n'));
+                    client.sendMessage(message.from, 'Last Chats: ' + parseHistoryDate[0].last_chatting + '\nStopped At: ' + parseHistoryDate[0].stopped_at + '\n\n' + chats.join('\n\n====================\n\n'));
                 }
             }
         }
